Refactor signal processing modules to use float instead of double for improved performance and consistency; update related validation functions and generator structures

+++ inc/disturbaces.h
@@ -27,7 +27,7 @@ struct DisturbanceGenerator
 DisturbanceStatus disturbance_init(
     struct DisturbanceGenerator *generator);
 
-double disturbance_generate(
+float disturbance_generate(
     struct DisturbanceGenerator *generator);
 
 #endif
